+++ src/Data/Graph/AdjacencyList/Network.hs
@@ -1,9 +1,9 @@
 {-|
 Module      : Data.Graph.AdjacencyList.Network
 Description : Flow network data type for max-flow problems
-Copyright   : Thodoris Papakonstantinou, 2017
-License     : GPL-3
-Maintainer  : mail@tpapak.com
+Copyright   : Thodoris Papakonstantinou, 2017-2026
+License     : LGPL-3
+Maintainer  : dev@tpapak.com
 Stability   : experimental
 Portability : POSIX